pyproject.toml
[project]
name = "minutes_of_meeting"
version = "0.1.0"
description = "minutes-of-meeting using crewAI"
authors = [{ name = "Arnav", email = "arnav.pathak2003@proton.me" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.102.0,<1.0.0",
    "google-api-python-client>=2.164.0",
    "google-auth>=2.38.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "markdown>=3.7",
    "openai-whisper>=20240930",
    "pydub>=0.25.1",
]

[project.scripts]
kickoff = "minutes_of_meeting.main:kickoff"
plot = "minutes_of_meeting.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
